refactor: cookies into its own state
