+++ src/util/mod.rs
@@ -1,2 +1,5 @@
+mod mutex_utils;
+pub(crate) use self::mutex_utils::*;
+
 mod new_random_socket_addr;
 pub use self::new_random_socket_addr::*;
